+++ src/container_inspector/image.py
@@ -356,14 +356,22 @@ class Image(ArchiveMixin, ConfigMixin):
         """
         return self.layers[0]
 
-    def extract_layers(self, extracted_location):
+    def extract_layers(self, extracted_location, skip_symlinks=True):
         """
         Extract all layer archives to the `extracted_location` directory.
         Each layer is extracted to its own directory named after its `layer_id`.
+        Skip symlinks and links if ``skip_symlinks`` is True.
+        Return a list of ExtractEvent.
         """
+        all_events = []
         for layer in self.layers:
             exloc = os.path.join(extracted_location, layer.layer_id)
-            layer.extract(extracted_location=exloc)
+            events = layer.extract(
+                extracted_location=exloc,
+                skip_symlinks=skip_symlinks,
+            )
+            all_events.extend(events)
+        return events
 
     def get_layers_resources(self, with_dir=False):
         """
@@ -453,9 +461,11 @@ class Image(ArchiveMixin, ConfigMixin):
     def extract(archive_location, extracted_location, skip_symlinks=False):
         """
         Extract the image archive tarball at ``archive_location`` to
-        ``extracted_location``. Skip symlinks and links if ``skip_symlinks`` is True.
+        ``extracted_location``.
+        Skip symlinks and links if ``skip_symlinks`` is True.
+        Return a list of ExtractEvent.
         """
-        utils.extract_tar(
+        return utils.extract_tar(
             location=archive_location,
             target_dir=extracted_location,
             skip_symlinks=skip_symlinks,
@@ -466,24 +476,29 @@ class Image(ArchiveMixin, ConfigMixin):
         archive_location,
         extracted_location,
         verify=True,
+        skip_symlinks=False,
     ):
         """
-        Return a list of Images found in the tarball at `archive_location` that
-        will be extracted to `extracted_location`. The tarball must be in the
+        Return a list of Images found in the tarball at ``archive_location`` that
+        will be extracted to ``extracted_location``. The tarball must be in the
         format of a "docker save" command tarball.
 
-        If `verify` is True, perform extra checks on the config data and layers
+        If ``verify`` is True, perform extra checks on the config data and layers
         checksums.
+        Skip symlinks and links if ``skip_symlinks`` is True.
+        Ignore the extract events from extraction.
         """
         if TRACE:
             logger.debug(
-                f'get_images_from_tarball: {archive_location} , '
+                f'get_images_from_tarball: {archive_location} '
                 f'extracting to: {extracted_location}'
             )
 
-        Image.extract(
+        # TODO: do not ignore extract events
+        _events = Image.extract(
             archive_location=archive_location,
             extracted_location=extracted_location,
+            skip_symlinks=skip_symlinks,
         )
 
         return Image.get_images_from_dir(
@@ -1071,13 +1086,15 @@ class Layer(ArchiveMixin, ConfigMixin):
         if not self.size:
             self.size = os.path.getsize(self.archive_location)
 
-    def extract(self, extracted_location, skip_symlinks=True):
+    def extract(self, extracted_location, skip_symlinks=False):
         """
         Extract this layer archive in the `extracted_location` directory and set
         this Layer ``extracted_location`` attribute to ``extracted_location``.
+        Skip symlinks and links if ``skip_symlinks`` is True.
+        Return a list of ExtractEvent.
         """
         self.extracted_location = extracted_location
-        utils.extract_tar(
+        return utils.extract_tar(
             location=self.archive_location,
             target_dir=extracted_location,
             skip_symlinks=skip_symlinks,